sql: fix overflow for soft limits
We're tracking `LimitHint` property as a float and later convert it to
an int64. Previously, this could lead to overflows when the limit hint
value exceeds the range of int64, and we now instead will use a value of
0 (i.e. disable the soft limit).

Release note (bug fix): Previously, CockroachDB could encounter an
internal error when evaluating queries with OFFSET and LIMIT clauses
when the addition of the `offset` and the `limit` value would be larger
than `int64` range.

+++ pkg/sql/distsql_spec_exec_factory.go
@@ -646,7 +646,7 @@ func (e *distSQLSpecExecFactory) ConstructIndexJoin(
 	keyCols []exec.NodeColumnOrdinal,
 	tableCols exec.TableColumnOrdinalSet,
 	reqOrdering exec.OutputOrdering,
-	limitHint int,
+	limitHint int64,
 ) (exec.Node, error) {
 	return nil, unimplemented.NewWithIssue(47473, "experimental opt-driven distsql planning: index join")
 }
@@ -666,7 +666,7 @@ func (e *distSQLSpecExecFactory) ConstructLookupJoin(
 	isSecondJoinInPairedJoiner bool,
 	reqOrdering exec.OutputOrdering,
 	locking *tree.LockingItem,
-	limitHint int,
+	limitHint int64,
 ) (exec.Node, error) {
 	// TODO (rohany): Implement production of system columns by the underlying scan here.
 	return nil, unimplemented.NewWithIssue(47473, "experimental opt-driven distsql planning: lookup join")
